backend/requirements.txt
pytest~=7.3.2
Faker~=12.0.1
djangorestframework~=3.14.0
Django~=4.2.2
djoser~=2.2.0

django-filter~=23.2
reportlab~=4.0.4